% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrapi.R
\name{get_leagues}
\alias{get_leagues}
\title{Get SoFIFA IDs for leagues}
\usage{
get_leagues()
}
\value{
A data frame with columns \code{league_id}, \code{league_name}
}
\description{
Get SoFIFA IDs for leagues
}
\examples{
# get all available leagues
leagues <- get_leagues()

# get all Premier League teams
teams <- get_league_teams(13)

# get all Tottenham Hotspur players
players <- get_team_players(18)
}
