% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience.R
\name{get_player_scores}
\alias{get_player_scores}
\alias{get_team_scores}
\alias{get_league_scores}
\title{Get SoFIFA scores for players, teams, or complete leagues.}
\usage{
get_player_scores(player_ids, max_results = Inf)

get_team_scores(team_id, max_results = Inf, include_on_loan = TRUE)

get_league_scores(league_id, max_results = Inf, include_on_loan = TRUE)
}
\arguments{
\item{player_ids}{Numeric (vector of) ID(s) obtained via \code{get_players()}.}

\item{max_results}{Numeric maximum results returned. Defaults to \code{Inf}.}

\item{team_id}{Numeric ID obtained via \code{get_teams()}.}

\item{include_on_loan}{Logical indicating if players actually on loan should
be included. Defaults to TRUE.}

\item{league_id}{Numeric ID obtained via \code{get_leagues()}.}
}
\value{
\code{player_scores}, \code{team_scores}, \code{league_scores},
  a data frame.
}
\description{
Get SoFIFA scores for players, teams, or complete leagues.
}
\examples{
\dontrun{
# get scores for Harry Kane
player_scores <- get_player_scores(202126)

# get scores for Harry Kane and Dele Alli
player_scores <- get_player_scores(c(202126, 211117))

# get scores for all Tottenham Hotspur players
team_scores <- get_team_scores(18, max_results = 5)

# get scores for all Premier League players
league_scores <- get_league_scores(13, max_results = 5)
}
}
