% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrapi.R
\name{get_players}
\alias{get_players}
\alias{get_leagues}
\alias{get_teams}
\alias{get_players}
\title{Get SoFIFA IDs for leagues, teams, or players.}
\usage{
get_leagues()

get_teams(league_id)

get_players(team_id)
}
\arguments{
\item{league_id}{Numeric ID obtained via \code{get_leagues()}.}

\item{team_id}{Numeric ID obtained via \code{get_teams()}.}
}
\value{
\code{leagues}, \code{teams}, \code{players},
  a data frame.
}
\description{
Get SoFIFA IDs for leagues, teams, or players.
}
\examples{
# get all available leagues
leagues <- get_leagues()

# get all Premier League teams
teams <- get_teams(13)

# get all Tottenham Hotspur players
players <- get_players(18)
}
