% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience.R
\name{get_player_scores}
\alias{get_player_scores}
\alias{get_team_scores}
\alias{get_league_scores}
\title{Get SoFIFA scores for players, teams, or complete leagues.}
\usage{
get_player_scores(player_ids)

get_team_scores(team_id)

get_league_scores(league_id)
}
\arguments{
\item{player_ids}{Numeric (vector of) ID(s) obtained via \code{get_players()}.}

\item{team_id}{Numeric ID obtained via \code{get_teams()}.}

\item{league_id}{Numeric ID obtained via \code{get_leagues()}.}
}
\value{
\code{player_scores}, \code{team_scores}, \code{league_scores},
  a data frame.
}
\description{
Get SoFIFA scores for players, teams, or complete leagues.
}
\examples{
# get scores for Harry Kane
player_scores <- get_player_scores(202126)

# get scores for Harry Kane and Dele Alli
player_scores <- get_player_scores(c(202126, 211117))

# get scores for all Tottenham Hotspur players
team_scores <- get_team_scores(18)

\dontrun{
# get scores for all Premier League players
league_scores <- get_league_scores(13)
}
}
