% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience.R
\name{get_player_scores}
\alias{get_player_scores}
\alias{get_player_history}
\alias{get_team_players}
\title{Get SoFIFA scores for players, teams, or complete leagues.}
\usage{
get_player_scores(player_ids, version = NULL, exportdate = NULL)

get_player_history(player_ids)

get_team_players(team_ids, include.onloan = FALSE)
}
\arguments{
\item{player_ids}{Numeric (vector of) ID(s) obtained via \code{get_players()}.}

\item{team_id}{Numeric ID obtained via \code{get_teams()}.}

\item{league_id}{Numeric ID obtained via \code{get_leagues()}.}

\item{max_results}{Numeric maximum results returned. Defaults to \code{Inf}.}
}
\value{
\code{player_scores}, \code{team_scores}, \code{league_scores},
  a data frame.
}
\description{
Get SoFIFA scores for players, teams, or complete leagues.
}
\examples{
# get scores for Harry Kane
player_scores <- get_player_scores(202126)

# get scores for Harry Kane and Dele Alli
player_scores <- get_player_scores(c(202126, 211117))

# get scores for all Tottenham Hotspur players
team_scores <- get_player_scores(head(get_team_players(18)$player_id, 5))

# get scores for all Premier League players
league_scores <- get_player_scores(head(get_team_players(get_league_teams(13)$team_id)$player_id, 5))
}
